requirements.txt
# Core Data & Computation
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0

# Financial Data
yfinance>=0.2.28
pandas-datareader>=0.10.0
fredapi>=0.5.0

# Portfolio Optimization
cvxpy>=1.3.0
# pypfopt>=1.5.0  # 移除，不再需要

# Machine Learning & LLM
anthropic>=0.25.0

# Data Processing
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Testing
pytest>=7.4.0